new examples

+++ src/main/java/io/millesabords/r2dbc/step3/repository/MovieRepository.java
@@ -4,7 +4,7 @@ import io.millesabords.r2dbc.step3.entity.Movie;
 import org.springframework.data.repository.reactive.ReactiveCrudRepository;
 import reactor.core.publisher.Mono;
 
-public interface MovieRepository extends ReactiveCrudRepository<Movie, Long> {
+public interface MovieRepository extends ReactiveCrudRepository<Movie, Integer> {
 
     Mono<Movie> findByTitle(String title);
 }
